pyproject.toml
[tool.poetry]
name = "megano shop"
version = "0.1.0"
description = "Online shop"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
django-mptt = "^0.14.0"
django-mptt-admin = "^2.3.0"
pillow = "^9.3.0"
django-mptt-urls = "^2.0.4"
django-cleanup = "^6.0.0"
django-constance = "^2.9.1"
django-picklefield = "^3.1"
django-extensions = "^3.2.1"
random-username = "^1.0.2"
docker-compose = "^1.29.2"
redis = "^4.5.1"
celery = "^5.2.7"
gunicorn = "^20.1.0"
fastapi = "^0.92.0"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

